@@ -5,17 +5,17 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>3.0.5</version>
+        <version>3.1.4</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>fr.les-projets-cagnottes</groupId>
     <artifactId>core</artifactId>
-    <version>0.16.3</version>
+    <version>0.17.0</version>
     <name>Les Projets Cagnottes - Core</name>
     <description>Les Projets Cagnottes - Main API Component</description>
     <properties>
-        <java.version>19</java.version>
-        <testcontainers.version>1.17.6</testcontainers.version>
+        <java.version>21</java.version>
+        <testcontainers.version>1.19.0</testcontainers.version>
     </properties>
     <dependencies>
         <dependency>
